Gate keeper added to controller.

+++ GoodSamaritan/Controllers/ClientController.cs
@@ -18,6 +18,11 @@ namespace GoodSamaritan.Controllers
         // GET: Client
         public async Task<ActionResult> Index()
         {
+            if (!User.IsInRole("Worker") && !User.IsInRole("Administrator"))
+            {
+                return null;
+            }
+
             var clientModel = db.ClientModel.Include(c => c.AbuserRealtionship).Include(c => c.Age).Include(c => c.AssignedWorker).Include(c => c.Crisis).Include(c => c.DuplicateFile).Include(c => c.Ethnicity).Include(c => c.FamilyViolenceFile).Include(c => c.FiscalYear).Include(c => c.Incident).Include(c => c.Program).Include(c => c.ReferralSource).Include(c => c.RepeatClient).Include(c => c.RiskLevel).Include(c => c.RiskStatus).Include(c => c.Service).Include(c => c.StatusOfFile).Include(c => c.VictimOfIncident);
             return View(await clientModel.ToListAsync());
         }
@@ -25,6 +30,10 @@ namespace GoodSamaritan.Controllers
         // GET: Client/Details/5
         public async Task<ActionResult> Details(int? id)
         {
+            if (!User.IsInRole("Worker") && !User.IsInRole("Administrator"))
+            {
+                return null;
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -40,6 +49,10 @@ namespace GoodSamaritan.Controllers
         // GET: Client/Create
         public ActionResult Create()
         {
+            if (!User.IsInRole("Worker") && !User.IsInRole("Administrator"))
+            {
+                return null;
+            }
             ViewBag.AbuserRelationshipId = new SelectList(db.AbuserRelationshipModel, "AbuserRelationshipId", "AbuserRelationship");
             ViewBag.AgeId = new SelectList(db.AgeModel, "AgeId", "Age");
             ViewBag.AssignedWorkerId = new SelectList(db.AssignedWorkerModel, "AssignedWorkerId", "AssignedWorker");
@@ -67,6 +80,10 @@ namespace GoodSamaritan.Controllers
         [ValidateAntiForgeryToken]
         public async Task<ActionResult> Create([Bind(Include = "ClientReferenceNumber,FiscalYearId,Month,Day,Surname,FirstName,PoliceFileNumber,CourtFileNumber,SwcFileNumber,RiskLevelId,CrisisId,ServiceId,ProgramId,RiskAssessmentAssignedTo,RiskStatusId,AssignedWorkerId,ReferralSourceId,IncidentId,AbuserSurnameName,AbuserFirstName,AbuserRelationshipId,VictimOfIncidentId,FamilyViolenceFileId,EthnicityId,AgeId,RepeatClientId,DuplicateFileId,NumberOfChildren0To6,NumberOfChildren7To12,NumberOfChildren13To18,StatusOfFileId,DateLastTransferred,DateClosed,DateReOpened")] ClientModel clientModel)
         {
+            if (!User.IsInRole("Worker") && !User.IsInRole("Administrator"))
+            {
+                return null;
+            }
             if (ModelState.IsValid)
             {
                 db.ClientModel.Add(clientModel);
@@ -97,6 +114,10 @@ namespace GoodSamaritan.Controllers
         // GET: Client/Edit/5
         public async Task<ActionResult> Edit(int? id)
         {
+            if (!User.IsInRole("Worker") && !User.IsInRole("Administrator"))
+            {
+                return null;
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -133,6 +154,10 @@ namespace GoodSamaritan.Controllers
         [ValidateAntiForgeryToken]
         public async Task<ActionResult> Edit([Bind(Include = "ClientReferenceNumber,FiscalYearId,Month,Day,Surname,FirstName,PoliceFileNumber,CourtFileNumber,SwcFileNumber,RiskLevelId,CrisisId,ServiceId,ProgramId,RiskAssessmentAssignedTo,RiskStatusId,AssignedWorkerId,ReferralSourceId,IncidentId,AbuserSurnameName,AbuserFirstName,AbuserRelationshipId,VictimOfIncidentId,FamilyViolenceFileId,EthnicityId,AgeId,RepeatClientId,DuplicateFileId,NumberOfChildren0To6,NumberOfChildren7To12,NumberOfChildren13To18,StatusOfFileId,DateLastTransferred,DateClosed,DateReOpened")] ClientModel clientModel)
         {
+            if (!User.IsInRole("Worker") && !User.IsInRole("Administrator"))
+            {
+                return null;
+            }
             if (ModelState.IsValid)
             {
                 db.Entry(clientModel).State = EntityState.Modified;
@@ -162,6 +187,10 @@ namespace GoodSamaritan.Controllers
         // GET: Client/Delete/5
         public async Task<ActionResult> Delete(int? id)
         {
+            if (!User.IsInRole("Worker") && !User.IsInRole("Administrator"))
+            {
+                return null;
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -179,6 +208,10 @@ namespace GoodSamaritan.Controllers
         [ValidateAntiForgeryToken]
         public async Task<ActionResult> DeleteConfirmed(int id)
         {
+            if (!User.IsInRole("Worker") && !User.IsInRole("Administrator"))
+            {
+                return null;
+            }
             ClientModel clientModel = await db.ClientModel.FindAsync(id);
             db.ClientModel.Remove(clientModel);
             await db.SaveChangesAsync();
